requirements.txt
streamlit
google-genai
nest_asyncio
Pillow